changing the connection time values to long from int

+++ wavemaker-app-runtime-core/src/main/java/com/wavemaker/runtime/soap/SoapServiceSettings.java
@@ -29,8 +29,8 @@ public class SoapServiceSettings {
 
     private String endpointAddress;
 
-    private int connectionTimeout;
-    private int requestTimeout;
+    private long connectionTimeout;
+    private long requestTimeout;
 
     private String soapActionURI;
 
@@ -95,7 +95,7 @@ public class SoapServiceSettings {
      *
      * @return The connection timeout value
      */
-    public int getConnectionTimeout() {
+    public long getConnectionTimeout() {
         return this.connectionTimeout;
     }
 
@@ -104,7 +104,7 @@ public class SoapServiceSettings {
      *
      * @param connectionTimeout The connection timeout value to set.
      */
-    public void setConnectionTimeout(int connectionTimeout) {
+    public void setConnectionTimeout(long connectionTimeout) {
         this.connectionTimeout = connectionTimeout;
     }
 
@@ -113,7 +113,7 @@ public class SoapServiceSettings {
      *
      * @return The request timeout value.
      */
-    public int getRequestTimeout() {
+    public long getRequestTimeout() {
         return this.requestTimeout;
     }
 
@@ -122,7 +122,7 @@ public class SoapServiceSettings {
      *
      * @param requestTimeout The request timeout value to set.
      */
-    public void setRequestTimeout(int requestTimeout) {
+    public void setRequestTimeout(long requestTimeout) {
         this.requestTimeout = requestTimeout;
     }
